% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSRRR.R
\name{iSRRR}
\alias{iSRRR}
\title{iSRRR: Integrative Sparse Reduced Rank Regression}
\usage{
iSRRR(
  X,
  Y,
  pvec,
  nrank,
  cutoff,
  params = NULL,
  control = NULL,
  trueB = NULL,
  use.gglasso = TRUE
)
}
\arguments{
\item{X}{An \code{n} by \code{p} design matrix.}

\item{Y}{An \code{n} by \code{q} response matrix.}

\item{pvec}{A numeric vector of the number of variable for each dataset.}

\item{nrank}{The rank of matrices to be estimated.}

\item{cutoff}{Hard-thresholding parameter. Default is 0 (no thresholding).}

\item{params}{A list with \code{lambda.seq}=NULL, \code{nlambda}=5, \code{lambda.factor}=1e-4, \code{log.scale}=TRUE, and \code{group.size}=FALSE.}

\item{control}{A list with \code{best}=FALSE, \code{early.stop}=TRUE, \code{rot.method}="quartimax" \code{maxit.B}=3e8, \code{eps.B}=1e-8, \code{maxit.mse}=50, \code{eps.mse}=1e-6, \code{X.scale}=c("group", "each"), \code{Y.scale}=FALSE, \code{verbose}=FALSE, and \code{threads}=1.}

\item{trueB}{A list contains X with dimension \code{n} by \code{p} and Y with dimension \code{n} by \code{q}.}

\item{use.gglasso}{Default is TRUE.}
}
\value{
A list with output objects
}
\description{
The iSRRR solves the constrained optimization problem with sparse regularization:

\eqn{ \frac{1}{2n} \| Y - X B A^T \|_F^2 +  \lambda \sum \sum \| b_{ik} \|_2}  subject to \eqn{A \in \mathcal{T}(v), ~ A \in \mathcal{O}_s(q,r) }
}
\references{
No reference.
}
\author{
Kipoong Kim <statpng@snu.ac.kr>, Sungkyu Jung <sungkyu@snu.ac.kr>
}
