% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{simdata}
\alias{simdata}
\alias{CorrAR}
\title{Simulation Data}
\usage{
CorrAR(p, rho)
}
\arguments{
\item{typeA}{One of \code{sparse0.1}, \code{sparse0.2}, \code{sparse0.3}, \code{sparse0.4}, \code{stiefel} and \code{perfect-simple}.}

\item{typeB}{One of \code{null}, \code{row}, \code{block}, \code{individual}, \code{joint}, \code{both}, \code{partial} and \code{all}.}

\item{n}{The number of samples. Default is 50.}

\item{d}{The number of multi-source datasets. Default is 3.}

\item{q}{The number of responses. Default is 10.}

\item{pvec}{A vector representing the number of variables for each dataset. Default is \code{rep(10, 3)}.}

\item{rvec}{A vector representing the rank of each dataset. Default is \code{rep(2, 3)}.}

\item{es}{Effect size: choose one of "1", "2", ..., "5" and "norm". Default is "5".}

\item{simplify}{A logical value for simplifying the matrix of \code{A} using the quartimax rotation. Default is TRUE.}

\item{snr}{A numeric value for SNR (signal to noise). Default is 1.}

\item{sigma}{A numeric value of a multiplying constant on the variance of an error term. Default is NULL. If NULL, sigma = |Y0|_F^2 / |E|_F^2) / sqrt(snr), where Y0 = XBA^T.}

\item{rho_X}{A numeric value for controlling the correlation between predictors.}

\item{rho_E}{A numeric value for controlling the correlation between errors.}
}
\value{
A list with Y, X, C, A, B, sigma, Xsigma, n, d, q, p, rvec, sim.params
}
\description{
Simulation Data with a sparse structural \code{B} and a sparse \code{A} with orthonormal columns.
}
\examples{

data <- simdata2(typeA="quartimax", typeB="individual", 
                 n=50, d=4, q=60, p=100, rvec=rep(2,4), 
                 snr=0.5)

data$B |> round(3)
data$A |> round(3)


}
\references{
No reference.
}
\author{
Kipoong Kim <statpng@snu.ac.kr>
}
