% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSRRR.R
\name{iSRRR}
\alias{iSRRR}
\title{iSRRR: Integrative Sparse Reduced Rank Regression}
\usage{
iSRRR(
  X,
  Y,
  pvec,
  nrank,
  nu,
  params = NULL,
  control = NULL,
  trueA = NULL,
  trueB = NULL
)
}
\arguments{
\item{X}{An \code{n} by \code{p} design matrix.}

\item{Y}{An \code{n} by \code{q} response matrix.}

\item{pvec}{A numeric vector of the number of variable for each dataset.}

\item{nrank}{The rank of matrices to be estimated.}

\item{nu}{Row-wise sparsity level on loading matrix A.}

\item{params}{A list with \code{lambda.seq}=NULL, \code{nlambda}=5, \code{lambda.factor}=1e-4, \code{log.scale}=TRUE, and \code{group.size}=TRUE.}

\item{control}{A list with \code{best}=FALSE, \code{rot.method}="quartimax" \code{maxit.B}=3e8, \code{eps.B}=1e-8, \code{maxit.mse}=50, \code{eps.mse}=1e-6, \code{X.scale}=c("group", "each"), \code{Y.scale}=FALSE, \code{verbose}=FALSE, and \code{threads}=1.}

\item{trueA}{A \code{q} x \code{r} dimensional matrix for the true parameter A. This is intended for simulation purpose.}

\item{trueB}{A \code{p} x \code{r} dimensional matrix for the true parameter B. This is intended for simulation purpose.}
}
\value{
A list with output objects along with nu values
}
\description{
The iSRRR solves the constrained optimization problem with sparse regularization:

\eqn{ \frac{1}{2n} \| Y - X B A^T \|_F^2 +  \lambda \sum \sum \| b_{ik} \|_2}  subject to \eqn{A \in \mathcal{T}(v), ~ A \in \mathcal{O}_s(q,r) }
}
\examples{
# test

}
\references{
Kipoong Kim and Sungkyu Jung. (2022). Integrative sparse reduced-rank regression via orthogonal rotation for analysis of high-dimensional multi-source data. Submitted.
}
\author{
Kipoong Kim <statpng@snu.ac.kr>, Sungkyu Jung <sungkyu@snu.ac.kr>
}
